Remove filename check (#12)
I'm removing a check that causes a rule to not be applied if filepath contains 'test'. Our CI uses a filepath like
'/export/home/tester/orca/MP_TRUNKDEV_DISTRIBUTED_TEST/workspace/foo/app.js', Because the word 'test' is in
the complete filepath, the rule is not applied to any files on CI.
Application of .eslint rules are expected to be handled via .eslintrc or .eslintignore.